requirements.txt
pandas
sentence-transformers
tqdm
faiss-cpu
numpy
pickle5
